requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docopt==0.6.2
flake8==6.0.0
idna==3.4
mccabe==0.7.0
Pillow==9.5.0
pipreqs==0.4.13
pycodestyle==2.10.0
pyflakes==3.0.1
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.0.3
yarg==0.1.9
